// vereinsmatching/src/lib/scouting/types.ts
export const POSITIONS = [
  'TW',
  'IV',
  'LV',
  'RV',
  'DM',
  'ZM',
  'OM',
  'LA',
  'RA',
  'ST',
] as const

export type Position = (typeof POSITIONS)[number]

export const POSITION_LABEL: Record<Position, string> = {
  TW: 'Torwart',
  IV: 'Innenverteidigung',
  LV: 'Linksverteidigung',
  RV: 'Rechtsverteidigung',
  DM: 'Defensives Mittelfeld',
  ZM: 'Zentrales Mittelfeld',
  OM: 'Offensives Mittelfeld',
  LA: 'Linksaußen',
  RA: 'Rechtsaußen',
  ST: 'Sturm',
}
// ...
export type Foot = 'links' | 'rechts' | 'beidfüßig'